基本情報B問題サンプル3問 データベース格納＋読み込み完了
