+++ src/app/(main)/issue_list/basic_info_b_problem/data/problem-logics.ts
@@ -1,0 +1,101 @@
+/**
+ * @file DBのlogicTypeと、対応するトレース関数をマッピングします。
+ */
+import type { VariablesState, TraceStep } from './problems'; // 既存の型定義を再利用
+
+// 各ロジックの定義
+const variableSwapLogic: { traceLogic: TraceStep[] } = {
+  traceLogic: [
+    (vars) => ({ ...vars, x: 1 }),
+    (vars) => ({ ...vars, y: 2 }),
+    (vars) => ({ ...vars, z: 3 }),
+    (vars) => ({ ...vars, x: vars.y }),
+    (vars) => ({ ...vars, y: vars.z }),
+    (vars) => ({ ...vars, z: vars.x }),
+    (vars) => vars,
+  ],
+};
+
+const fizzBuzzLogic: { traceLogic: TraceStep[]; calculateNextLine: (currentLine: number, vars: VariablesState) => number } = {
+  traceLogic: [
+    (vars) => vars,
+    (vars) => ({ ...vars, result: null }),
+    (vars) => vars,
+    (vars) => ({ ...vars, result: "3と5で割り切れる" }),
+    (vars) => vars,
+    (vars) => ({ ...vars, result: "3で割り切れる" }),
+    (vars) => vars,
+    (vars) => ({ ...vars, result: "5で割り切れる" }),
+    (vars) => vars,
+    (vars) => ({ ...vars, result: "3でも5でも割り切れない" }),
+    (vars) => vars,
+    (vars) => vars,
+  ],
+  calculateNextLine: (currentLine, vars) => {
+    const num = vars.num as number;
+    if (num === null) return currentLine;
+    switch (currentLine + 1) {
+      case 3: return num % 15 === 0 ? 3 : 4;
+      case 4: return 11;
+      case 5: return num % 3 === 0 ? 5 : 6;
+      case 6: return 11;
+      case 7: return num % 5 === 0 ? 7 : 8;
+      case 8: return 11;
+      case 9: return 9;
+      case 10: return 11;
+      default: return currentLine + 1;
+    }
+  },
+};
+
+const arraySumLogic: { traceLogic: TraceStep[]; calculateNextLine: (currentLine: number, vars: VariablesState) => number } = {
+    traceLogic: [
+        (vars) => ({ ...vars, in: [3, 2, 1, 6, 5, 4] }),
+        (vars) => ({ ...vars, out: [] }),
+        (vars) => ({ ...vars, i: null, tail: null }),
+        (vars) => {
+            const newOut = [...(vars.out as number[])];
+            newOut.push((vars.in as number[])[0]);
+            return { ...vars, out: newOut };
+        },
+        (vars) => {
+            const i = vars.i as number | null;
+            return i === null ? { ...vars, i: 2 } : vars;
+        },
+        (vars) => {
+            const out = vars.out as number[];
+            return { ...vars, tail: out[out.length - 1] };
+        },
+        (vars) => {
+            const newOut = [...(vars.out as number[])];
+            const i = vars.i as number;
+            const valueToAdd = (vars.tail as number) + (vars.in as number[])[i - 1];
+            newOut.push(valueToAdd);
+            return { ...vars, out: newOut };
+        },
+        (vars) => {
+            const i = vars.i as number;
+            return { ...vars, i: i + 1 };
+        },
+        (vars) => vars,
+    ],
+    calculateNextLine: (currentLine, vars) => {
+        const i = vars.i as number;
+        const inArray = vars.in as number[];
+        switch (currentLine + 1) {
+            case 5: return i <= inArray.length ? 5 : 8;
+            case 6: return 6;
+            case 7: return 7;
+            case 8: return i <= inArray.length ? 5 : 8;
+            default: return currentLine + 1;
+        }
+    },
+};
+
+
+// logicTypeをキーとして、対応するロジックを返すマップ
+export const problemLogicsMap = {
+  'VARIABLE_SWAP': variableSwapLogic,
+  'FIZZ_BUZZ': fizzBuzzLogic,
+  'ARRAY_SUM': arraySumLogic,
+};
