+++ src/prisma/seed.ts
@@ -4,7 +4,7 @@ import { Prisma, PrismaClient } from '@prisma/client';
 // 例: import bcrypt from 'bcrypt';
 
 // ★重要★ あなたの環境の `problems.ts` ファイルへのパスが正しいか確認してください
-import { problems as localProblems } from '../app/issue_list/basic_info_b_problem/data/problems'; 
+import { problems as localProblems } from '../app/(main)/issue_list/basic_info_b_problem/data/problems'; 
 
 // PrismaClientを初期化
 const prisma = new PrismaClient();
